personal_projects/32.book-reader/backend/requirements.txt
fastapi
uvicorn
elevenlabs
python-dotenv
